+++ src/armies/utils.ts
@@ -2,6 +2,8 @@ import {Kingdom} from '../kingdom/data/kingdom';
 import {saveKingdom} from '../kingdom/storage';
 import {getBooleanSetting} from '../settings';
 import {distinctBy, sum} from '../utils';
+import {Modifier} from '../kingdom/modifiers';
+import {isEffectItem} from '../camping/actor';
 
 function isArmyActor(actor: Actor | null | undefined): boolean {
     return actor?.type === 'army';
@@ -48,4 +50,75 @@ export function calculateTotalArmyConsumption(game: Game): number {
         .map(a => (a as unknown as ArmyActor).system.consumption ?? 0)
         .forEach(c => consumption.push(c));
     return sum(consumption);
+}
+
+function getEffectCount(actor: Actor, slug: string): number {
+    return (actor.items
+            .find(i => isEffectItem(i) && i.slug === slug) as EffectItem | undefined)
+            ?.badge
+            ?.value
+        ?? 0;
+}
+
+export function getSelectedArmies(game: Game): (Actor & ArmyActor) [] {
+    return Array.from(game.user?.targets ?? [])
+        .filter(t => isArmyActor(t.actor))
+        .map(t => t.actor) as (Actor & ArmyActor)[];
+}
+
+export function getArmyModifiers(game: Game): Modifier[] {
+    const result: Modifier[] = [];
+    const selectedArmies = getSelectedArmies(game);
+    if (selectedArmies.length > 0) {
+        const first = selectedArmies[0];
+        const miredCount = getEffectCount(first, 'mired');
+        const wearyCount = getEffectCount(first, 'weary');
+        const baseModifier: Pick<Modifier, 'type' | 'phases' | 'enabled'> = {
+            type: 'circumstance',
+            phases: ['army'],
+            enabled: true,
+        };
+        if (wearyCount > 0) {
+            result.push({
+                ...baseModifier,
+                value: -wearyCount,
+                name: 'Weary',
+            });
+        }
+        if (miredCount > 0) {
+            result.push({
+                ...baseModifier,
+                value: -miredCount,
+                name: 'Mired',
+                activities: ['deploy-army'],
+            });
+        }
+    }
+    return result;
+}
+
+export function getScoutingDC(game: Game): number {
+    return Math.max(...getSelectedArmies(game)
+        .map(a => a.system.scouting ?? 0), 0);
+}
+
+export function isCampaignFeature(item: Item<unknown>): item is CampaignFeaturePF2E {
+    return item.type === 'campaignFeature';
+}
+
+export function isArmyTactic(item: Item<unknown>): item is CampaignFeaturePF2E {
+    return isCampaignFeature(item)
+        && item.system.campaign === 'kingmaker'
+        && item.system.category === 'army-tactic';
+}
+
+function filterArmyTactics(items: Item[]): CampaignFeaturePF2E[] {
+    const campaignFeatures = (items?.filter(i => isCampaignFeature(i)) ?? []) as unknown as CampaignFeaturePF2E[];
+    return campaignFeatures.filter(i => isArmyTactic(i));
+}
+
+export async function getArmyTactics(game: Game): Promise<CampaignFeaturePF2E[]> {
+    const documents = (await game.packs.get('pf2e.kingmaker-features')?.getDocuments()) as Item[];
+    const items = (game.items ?? []) as Item[];
+    return filterArmyTactics([...documents, ...items]);
 }